+++ pkg/lint/lint_test.go
@@ -7,6 +7,7 @@ import (
 	"os"
 	"testing"
 
+	"github.com/redpanda-data/helm-charts/charts/redpanda"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"sigs.k8s.io/yaml"
@@ -74,5 +75,26 @@ func TestOperatorArtifactHubImages(t *testing.T) {
 		chart.Annotations["artifacthub.io/images"],
 		fmt.Sprintf("%s:%s", configuratorRepo, chart.AppVersion),
 		"artifacthub.io/images should be in sync with .appVersion",
+	)
+}
+
+func TestRedpandaControllersTag(t *testing.T) {
+	chartBytes, err := os.ReadFile("../../charts/operator/Chart.yaml")
+	require.NoError(t, err)
+
+	valuesYAML, err := os.ReadFile("../../charts/redpanda/values.yaml")
+	require.NoError(t, err)
+
+	var chart map[string]any
+	require.NoError(t, yaml.Unmarshal(chartBytes, &chart))
+
+	var values redpanda.Values
+	require.NoError(t, yaml.Unmarshal(valuesYAML, &values))
+
+	require.Equal(
+		t,
+		chart["appVersion"].(string),
+		string(values.Statefulset.SideCars.Controllers.Image.Tag),
+		"the redpanda chart's values.yaml's controllers tag should be equal to the operator chart's appVersion",
 	)
 }